Moved RtApi subclass declarations from RtAudio.h to RtAudio.cpp
